+++ Changelog.md
@@ -1,6 +1,8 @@
 # Unreleased
 
 * Fix: Use a better exception message for invalid length validator configurations
+* Feature: Add support for zxcvbn via https://github.com/techgaun/zxcvbn-elixir
+  * Adds zxcvbn as a depdendency
 
 # 0.2.1 (2019-01-23)